audit-log/requirements.txt
# requirements.txt v0.1.0 (2025-08-20)
psycopg2-binary>=2.9.9
prometheus-client>=0.20.0
opentelemetry-sdk>=1.24.0
python-dotenv>=1.0.1
pika>=1.3.2
